@@ -5,11 +5,12 @@
 package validator
 
 import (
-	"github.com/elastic/package-spec/code/go/internal/fspath"
 	"testing"
 
 	"github.com/Masterminds/semver/v3"
 	"github.com/stretchr/testify/require"
+
+	"github.com/elastic/package-spec/v2/code/go/internal/fspath"
 )
 
 func TestNewSpec(t *testing.T) {
@@ -40,7 +41,6 @@ func TestNoBetaFeatures_Package_GA(t *testing.T) {
 	s := Spec{
 		*semver.MustParse("1.0.0"),
 		fspath.DirFS("testdata/fakespec"),
-		"1",
 	}
 	pkg, err := NewPackage("testdata/packages/features_ga")
 	require.NoError(t, err)
@@ -54,12 +54,11 @@ func TestBetaFeatures_Package_GA(t *testing.T) {
 	s := Spec{
 		*semver.MustParse("1.0.0"),
 		fspath.DirFS("testdata/fakespec"),
-		"1",
 	}
 	pkg, err := NewPackage("testdata/packages/features_beta")
 	require.NoError(t, err)
 
 	errs := s.ValidatePackage(*pkg)
-	require.Len(t,errs, 1)
+	require.Len(t, errs, 1)
 	require.Equal(t, errs[0].Error(), "spec for [testdata/packages/features_beta/beta] defines beta features which can't be enabled for packages with a stable semantic version")
 }
